Migrate to flake.nix from shell.nix
This is not strictly necessary, but I'm on a Nix Flake learning binge
and thought I'd like to try.

First I found that I can't simply use shellHook to define aliases any
more, since I can't run them from `nix develop`. I have to create
"proper" derivations to run them with `nix run`.

The tricky one here was the `publish-blog` derivation: it could not be
a simple shell script, as it needed to pull in my custom Emacs.